smart-plug: Only turn off speakers in the network pre-down script if the system is shutting down

+++ smart_plug/linux/turn-off-speakers.bash
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-sudo systemctl stop speakers.service
+# Only stop the speakers if the system is being shut down
+if [[ $(systemctl is-system-running || true) == 'stopping' ]]; then
+  sudo systemctl stop speakers.service
+fi
